+++ packages/list/src/mwc-radio-list-item-base.ts
@@ -79,6 +79,6 @@ export class RadioListItemBase extends ListItemBase {
     super.connectedCallback();
     this.addEventListener('click', this.boundOnClick);
 
-    this.toggleAttribute('mwc-list-item', true);
+    this.setAttribute('mwc-list-item', '');
   }
 }